scraper/requirements.txt
selenium
webdriver-manager
boto3
psycopg2-binary
python-dotenv
requests
urllib3